Extract and Refactor fake crud methods into reusable extensions DynamicsValue/fake-xrm-easy#122 DynamicsValue/fake-xrm-easy#70
