+++ src/FakeXrmEasy.Core/Middleware/Crud/MiddlewareBuilderExtensions.Crud.cs
@@ -57,13 +57,13 @@ namespace FakeXrmEasy.Middleware.Crud
                 #endif
                 
                 context.SetProperty(crudMessageExecutors);
-                AddFakeCreate(context, service);
-                AddFakeRetrieve(context, service);
-                AddFakeRetrieveMultiple(context, service);
-                AddFakeUpdate(context,service);
-                AddFakeDelete(context,service);
-                AddFakeAssociate(service);
-                AddFakeDisassociate(service);
+                service.AddFakeCreate()
+                    .AddFakeRetrieve()
+                    .AddFakeRetrieveMultiple()
+                    .AddFakeUpdate()
+                    .AddFakeDelete()
+                    .AddFakeAssociate()
+                    .AddFakeDisassociate();
             });
 
             return builder;
@@ -94,7 +94,6 @@ namespace FakeXrmEasy.Middleware.Crud
         /// <returns></returns>
         public static IMiddlewareBuilder UseCrud(this IMiddlewareBuilder builder) 
         {
-
             Func<OrganizationRequestDelegate, OrganizationRequestDelegate> middleware = next => {
 
                 return (IXrmFakedContext context, OrganizationRequest request) => {
@@ -133,108 +132,5 @@ namespace FakeXrmEasy.Middleware.Crud
             return fakeMessageExecutor.Execute(request, context);
         }
 
-        private static void AddFakeCreate(IXrmFakedContext context, IOrganizationService service) 
-        {
-            A.CallTo(() => service.Create(A<Entity>._))
-                .ReturnsLazily((Entity e) =>
-                {
-                    var request = new CreateRequest();
-                    request.Target = e;
-                    var response = service.Execute(request) as CreateResponse;
-                    return response.id;
-                });
-        }
-
-        private static void AddFakeUpdate(IXrmFakedContext context, IOrganizationService service) 
-        {
-            A.CallTo(() => service.Update(A<Entity>._))
-                .Invokes((Entity e) =>
-                {
-                    var request = new UpdateRequest();
-                    request.Target = e;
-                    service.Execute(request);
-                });
-        }
-
-        private static void AddFakeRetrieveMultiple(IXrmFakedContext context, IOrganizationService service)
-        {
-            //refactored from RetrieveMultipleExecutor
-            A.CallTo(() => service.RetrieveMultiple(A<QueryBase>._))
-                .ReturnsLazily((QueryBase req) => {
-                    var request = new RetrieveMultipleRequest { Query = req };
-
-                    var response = service.Execute(request) as RetrieveMultipleResponse;
-                    return response.EntityCollection;
-                });
-        }
-
-        private static void AddFakeRetrieve(IXrmFakedContext context, IOrganizationService service)
-        {
-            A.CallTo(() => service.Retrieve(A<string>._, A<Guid>._, A<ColumnSet>._))
-                .ReturnsLazily((string entityName, Guid id, ColumnSet columnSet) =>
-                {
-                    var request = new RetrieveRequest()
-                    {
-                        Target = new EntityReference() { LogicalName = entityName, Id = id },
-                        ColumnSet = columnSet
-                    };
-
-                    var response = service.Execute(request) as RetrieveResponse;
-                    return response.Entity;
-                });
-        }
-
-        private static void AddFakeDelete(IXrmFakedContext context, IOrganizationService service)
-        {
-            A.CallTo(() => service.Delete(A<string>._, A<Guid>._))
-                .Invokes((string entityName, Guid id) =>
-                {
-                    if (string.IsNullOrWhiteSpace(entityName))
-                    {
-                        throw new InvalidOperationException("The entity logical name must not be null or empty.");
-                    }
-
-                    if (id == Guid.Empty)
-                    {
-                        throw new InvalidOperationException("The id must not be empty.");
-                    }
-
-                    var entityReference = new EntityReference(entityName, id);
-
-                    var request = new DeleteRequest() { Target = entityReference };
-                    service.Execute(request);
-                });
-        }
-
-        private static void AddFakeAssociate(IOrganizationService service)
-        {
-            A.CallTo(() => service.Associate(A<string>._, A<Guid>._, A<Relationship>._, A<EntityReferenceCollection>._))
-                .Invokes((string entityName, Guid entityId, Relationship relationship, EntityReferenceCollection entityCollection) =>
-                {
-                    var request = new AssociateRequest()
-                    {
-                        Target = new EntityReference() { Id = entityId, LogicalName = entityName },
-                        Relationship = relationship,
-                        RelatedEntities = entityCollection
-                    };
-                    service.Execute(request);
-                });
-        }
-
-        private static void AddFakeDisassociate(IOrganizationService service)
-        {
-            A.CallTo(() => service.Disassociate(A<string>._, A<Guid>._, A<Relationship>._, A<EntityReferenceCollection>._))
-                .Invokes((string entityName, Guid entityId, Relationship relationship, EntityReferenceCollection entityCollection) =>
-                {
-                    var request = new DisassociateRequest()
-                    {
-                        Target = new EntityReference() { Id = entityId, LogicalName = entityName },
-                        Relationship = relationship,
-                        RelatedEntities = entityCollection
-                    };
-                    service.Execute(request);
-                });
-        }
-        
     }
 }
